+++ training/Utility/transactions.cs
@@ -12,7 +12,7 @@ namespace training.Utility
         public string sellerId { get; set; }
         public string costumerId { get; set; }
         public DateTime transactionDate { get; set; }
-        public int amount { get; set; }
+        public double amount { get; set; }
         public string status { get; set; }
 
 
@@ -21,7 +21,7 @@ namespace training.Utility
 
         }
 
-        public transactions(string sellerId, string costumerId, DateTime date, int amount, string status)
+        public transactions(string sellerId, string costumerId, DateTime date, double amount, string status)
         {
             _id = new ObjectId();
             this.sellerId = sellerId;
